Fix H2 codegen VARCHAR length issue and update AVG docs
- Filter out H2's default VARCHAR lengths (Int.MaxValue, 1000000000) in H2ColumnBuilder.length to prevent Derby execution failures during codegen round-trip testing
- Update SQL-to-Slick documentation examples to reflect AVG function returning Option[Double] instead of Option[Int]

+++ doc/code/SqlToSlick.scala
@@ -224,7 +224,7 @@ object SqlToSlick extends App {
             select ADDRESS_ID, AVG(AGE)
             from PERSON
             group by ADDRESS_ID
-          """.as[(Int,Option[Int])]
+          """.as[(Int,Option[Double])]
           //#sqlQueryGroupBy
         val slick =
           //#slickQueryGroupBy
@@ -235,7 +235,7 @@ object SqlToSlick extends App {
         val (sqlRes, slickRes) = Await.result(db.run(sql zip slick), Duration.Inf)
         assert(sqlRes == slickRes)
         assert(sqlRes.size > 0)
-        assert(sqlRes.exists(_._2 == Some(1000)))
+        assert(sqlRes.exists(_._2 == Some(1000.0)))
       };{
         val sql =
           //#sqlQueryHaving
